Switch from PostgreSQL to SQLite; add pm2 deployment guide
- app/config.py + .env: DATABASE_URL now defaults to sqlite:///./oddsapp.db, no external DB server needed
- app/db.py: enable check_same_thread=False and PRAGMA foreign_keys=ON per-connection (SQLite disables FK enforcement by default, which our cascade deletes rely on)
- models: primary keys changed from BigInteger to Integer — SQLite only aliases a column to its rowid (enabling autoincrement) when the column is declared as exactly INTEGER, so BigInteger silently broke autoincrement on insert
- alembic: rewrote the init migration's now() server_default to CURRENT_TIMESTAMP (portable across sqlite/postgres) and enabled render_as_batch for sqlite in env.py for future ALTER-style migrations
- tests/conftest.py: switched from a Postgres test database to an in-memory SQLite engine (StaticPool), so pytest needs no external setup at all
- removed psycopg2-binary dependency
- ecosystem.config.js + README rewrite: full git-clone-to-pm2 runbook (system deps, clone, pip install, alembic upgrade, pm2 start/save/startup), Caddy section kept, systemd kept as an alternative to pm2

+++ odds-app/app/models/combo.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from decimal import Decimal
 
-from sqlalchemy import BigInteger, DateTime, ForeignKey, Numeric, String, func
+from sqlalchemy import DateTime, ForeignKey, Integer, Numeric, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db import Base
@@ -10,16 +10,16 @@ from app.db import Base
 class ComboRecommendation(Base):
     __tablename__ = "combo_recommendations"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     fixture_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("fixtures.id", ondelete="CASCADE"), nullable=False
+        Integer, ForeignKey("fixtures.id", ondelete="CASCADE"), nullable=False
     )
     leg_a_market_id: Mapped[int | None] = mapped_column(
-        BigInteger, ForeignKey("markets.id"), nullable=True
+        Integer, ForeignKey("markets.id"), nullable=True
     )
     leg_a_selection: Mapped[str | None] = mapped_column(String(50), nullable=True)
     leg_b_market_id: Mapped[int | None] = mapped_column(
-        BigInteger, ForeignKey("markets.id"), nullable=True
+        Integer, ForeignKey("markets.id"), nullable=True
     )
     leg_b_selection: Mapped[str | None] = mapped_column(String(50), nullable=True)
     combo_type: Mapped[str | None] = mapped_column(String(50), nullable=True)
