+++ odds-app/app/models/odds.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from decimal import Decimal
 
-from sqlalchemy import BigInteger, DateTime, ForeignKey, Numeric, String, func
+from sqlalchemy import DateTime, ForeignKey, Integer, Numeric, String, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db import Base
@@ -10,9 +10,9 @@ from app.db import Base
 class Odds(Base):
     __tablename__ = "odds"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     market_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("markets.id", ondelete="CASCADE"), nullable=False
+        Integer, ForeignKey("markets.id", ondelete="CASCADE"), nullable=False
     )
     selection: Mapped[str] = mapped_column(String(50), nullable=False)
     decimal_odds: Mapped[Decimal] = mapped_column(Numeric(6, 3), nullable=False)
